Refactoring and improving documentation
