+++ src/Mod/Draft/drafttaskpanels/task_circulararray.py
@@ -25,16 +25,13 @@
 # \ingroup DRAFT
 # \brief This module provides the task panel code for the CircularArray tool.
 
-import PySide.QtGui as QtGui
-from PySide.QtCore import QT_TRANSLATE_NOOP
-
 import FreeCAD as App
 import FreeCADGui as Gui
 import Draft_rc  # include resources, icons, ui files
 import DraftVecUtils
-import draftutils.utils as utils
 
 from FreeCAD import Units as U
+from drafttaskpanels.task_base import TaskPanelPolarCircularBase
 from draftutils.messages import _msg, _wrn, _err, _log
 from draftutils.translate import _tr
 
@@ -42,7 +39,7 @@ from draftutils.translate import _tr
 bool(Draft_rc.__name__)
 
 
-class TaskPanelCircularArray:
+class TaskPanelCircularArray(TaskPanelPolarCircularBase):
     """TaskPanel code for the CircularArray command.
 
     The names of the widgets are defined in the `.ui` file.
@@ -83,16 +80,7 @@ class TaskPanelCircularArray:
         # The .ui file must be loaded into an attribute
         # called `self.form` so that it is displayed in the task panel.
         ui_file = ":/ui/TaskPanel_CircularArray.ui"
-        self.form = Gui.PySideUic.loadUi(ui_file)
-
         icon_name = "Draft_CircularArray"
-        svg = ":/icons/" + icon_name
-        pix = QtGui.QPixmap(svg)
-        icon = QtGui.QIcon.fromTheme(icon_name, QtGui.QIcon(svg))
-        self.form.setWindowIcon(icon)
-        self.form.setWindowTitle(_tr(self.name))
-
-        self.form.label_icon.setPixmap(pix.scaled(32, 32))
 
         # -------------------------------------------------------------------
         # Default values for the internal function,
@@ -102,6 +90,8 @@ class TaskPanelCircularArray:
 
         self.r_distance = 2 * start_distance.Value
         self.tan_distance = start_distance.Value
+
+        super(TaskPanelCircularArray, self).__init__(ui_file, icon_name)
 
         self.form.spinbox_r_distance.setProperty('rawValue',
                                                  self.r_distance)
@@ -119,63 +109,6 @@ class TaskPanelCircularArray:
         # TODO: the axis is currently fixed, it should be editable
         # or selectable from the task panel
         self.axis = App.Vector(0, 0, 1)
-
-        start_point = U.Quantity(0.0, App.Units.Length)
-        length_unit = start_point.getUserPreferred()[2]
-
-        self.center = App.Vector(start_point.Value,
-                                 start_point.Value,
-                                 start_point.Value)
-
-        self.form.input_c_x.setProperty('rawValue', self.center.x)
-        self.form.input_c_x.setProperty('unit', length_unit)
-        self.form.input_c_y.setProperty('rawValue', self.center.y)
-        self.form.input_c_y.setProperty('unit', length_unit)
-        self.form.input_c_z.setProperty('rawValue', self.center.z)
-        self.form.input_c_z.setProperty('unit', length_unit)
-
-        self.fuse = utils.get_param("Draft_array_fuse", False)
-        self.use_link = utils.get_param("Draft_array_Link", True)
-
-        self.form.checkbox_fuse.setChecked(self.fuse)
-        self.form.checkbox_link.setChecked(self.use_link)
-        # -------------------------------------------------------------------
-
-        # Some objects need to be selected before we can execute the function.
-        self.selection = None
-
-        # This is used to test the input of the internal function.
-        # It should be changed to True before we can execute the function.
-        self.valid_input = False
-
-        self.set_widget_callbacks()
-
-        self.tr_true = QT_TRANSLATE_NOOP("Draft", "True")
-        self.tr_false = QT_TRANSLATE_NOOP("Draft", "False")
-
-        # The mask is not used at the moment, but could be used in the future
-        # by a callback to restrict the coordinates of the pointer.
-        self.mask = ""
-
-    def set_widget_callbacks(self):
-        """Set up the callbacks (slots) for the widget signals."""
-        # New style for Qt5
-        self.form.button_reset.clicked.connect(self.reset_point)
-
-        # When the checkbox changes, change the internal value
-        self.form.checkbox_fuse.stateChanged.connect(self.set_fuse)
-        self.form.checkbox_link.stateChanged.connect(self.set_link)
-
-        # Old style for Qt4, avoid!
-        # QtCore.QObject.connect(self.form.button_reset,
-        #                        QtCore.SIGNAL("clicked()"),
-        #                        self.reset_point)
-        # QtCore.QObject.connect(self.form.checkbox_fuse,
-        #                        QtCore.SIGNAL("stateChanged(int)"),
-        #                        self.set_fuse)
-        # QtCore.QObject.connect(self.form.checkbox_link,
-        #                        QtCore.SIGNAL("stateChanged(int)"),
-        #                        self.set_link)
 
     def accept(self):
         """Execute when clicking the OK button or Enter key."""
@@ -313,16 +246,6 @@ class TaskPanelCircularArray:
         symmetry = self.form.spinbox_symmetry.value()
         return number, symmetry
 
-    def get_center(self):
-        """Get the value of the center from the widgets."""
-        c_x_str = self.form.input_c_x.text()
-        c_y_str = self.form.input_c_y.text()
-        c_z_str = self.form.input_c_z.text()
-        center = App.Vector(U.Quantity(c_x_str).Value,
-                            U.Quantity(c_y_str).Value,
-                            U.Quantity(c_z_str).Value)
-        return center
-
     def get_axis(self):
         """Get the axis that will be used for the array. NOT IMPLEMENTED.
 
@@ -330,46 +253,6 @@ class TaskPanelCircularArray:
         as an axis.
         """
         return self.axis
-
-    def reset_point(self):
-        """Reset the center point to the original distance."""
-        self.form.input_c_x.setProperty('rawValue', 0)
-        self.form.input_c_y.setProperty('rawValue', 0)
-        self.form.input_c_z.setProperty('rawValue', 0)
-
-        self.center = self.get_center()
-        _msg(_tr("Center reset:")
-             + " ({0}, {1}, {2})".format(self.center.x,
-                                         self.center.y,
-                                         self.center.z))
-
-    def print_fuse_state(self, fuse):
-        """Print the fuse state translated."""
-        if fuse:
-            state = self.tr_true
-        else:
-            state = self.tr_false
-        _msg(_tr("Fuse:") + " {}".format(state))
-
-    def set_fuse(self):
-        """Execute as a callback when the fuse checkbox changes."""
-        self.fuse = self.form.checkbox_fuse.isChecked()
-        self.print_fuse_state(self.fuse)
-        utils.set_param("Draft_array_fuse", self.fuse)
-
-    def print_link_state(self, use_link):
-        """Print the link state translated."""
-        if use_link:
-            state = self.tr_true
-        else:
-            state = self.tr_false
-        _msg(_tr("Create Link array:") + " {}".format(state))
-
-    def set_link(self):
-        """Execute as a callback when the link checkbox changes."""
-        self.use_link = self.form.checkbox_link.isChecked()
-        self.print_link_state(self.use_link)
-        utils.set_param("Draft_array_Link", self.use_link)
 
     def print_messages(self):
         """Print messages about the operation."""
@@ -391,114 +274,3 @@ class TaskPanelCircularArray:
                                          self.center.z))
         self.print_fuse_state(self.fuse)
         self.print_link_state(self.use_link)
-
-    def display_point(self, point=None, plane=None, mask=None):
-        """Display the coordinates in the x, y, and z widgets.
-
-        This function should be used in a Coin callback so that
-        the coordinate values are automatically updated when the
-        mouse pointer moves.
-        This was copied from `DraftGui.py` but needs to be improved
-        for this particular command.
-
-        point: Base::Vector3
-            is a vector that arrives by the callback.
-        plane: WorkingPlane
-            is a `WorkingPlane` instance, for example,
-            `App.DraftWorkingPlane`. It is not used at the moment,
-            but could be used to set up the grid.
-        mask: str
-            is a string that specifies which coordinate is being
-            edited. It is used to restrict edition of a single coordinate.
-            It is not used at the moment but could be used with a callback.
-        """
-        # Get the coordinates to display
-        d_p = None
-        if point:
-            d_p = point
-
-        # Set the widgets to the value of the mouse pointer.
-        #
-        # setProperty() is used if the widget is a FreeCAD widget like
-        # Gui::InputField or Gui::QuantitySpinBox, which are based on
-        # QLineEdit and QAbstractSpinBox.
-        #
-        # setText() is used to set the text inside the widget, this may be
-        # useful in some circumstances.
-        #
-        # The mask allows editing only one field, that is, only one coordinate.
-        # sbx = self.form.spinbox_c_x
-        # sby = self.form.spinbox_c_y
-        # sbz = self.form.spinbox_c_z
-        if d_p:
-            if self.mask in ('y', 'z'):
-                # sbx.setText(displayExternal(d_p.x, None, 'Length'))
-                self.form.input_c_x.setProperty('rawValue', d_p.x)
-            else:
-                # sbx.setText(displayExternal(d_p.x, None, 'Length'))
-                self.form.input_c_x.setProperty('rawValue', d_p.x)
-            if self.mask in ('x', 'z'):
-                # sby.setText(displayExternal(d_p.y, None, 'Length'))
-                self.form.input_c_y.setProperty('rawValue', d_p.y)
-            else:
-                # sby.setText(displayExternal(d_p.y, None, 'Length'))
-                self.form.input_c_y.setProperty('rawValue', d_p.y)
-            if self.mask in ('x', 'y'):
-                # sbz.setText(displayExternal(d_p.z, None, 'Length'))
-                self.form.input_c_z.setProperty('rawValue', d_p.z)
-            else:
-                # sbz.setText(displayExternal(d_p.z, None, 'Length'))
-                self.form.input_c_z.setProperty('rawValue', d_p.z)
-
-        if plane:
-            pass
-
-        # Set masks
-        if (mask == "x") or (self.mask == "x"):
-            self.form.input_c_x.setEnabled(True)
-            self.form.input_c_y.setEnabled(False)
-            self.form.input_c_z.setEnabled(False)
-            self.set_focus("x")
-        elif (mask == "y") or (self.mask == "y"):
-            self.form.input_c_x.setEnabled(False)
-            self.form.input_c_y.setEnabled(True)
-            self.form.input_c_z.setEnabled(False)
-            self.set_focus("y")
-        elif (mask == "z") or (self.mask == "z"):
-            self.form.input_c_x.setEnabled(False)
-            self.form.input_c_y.setEnabled(False)
-            self.form.input_c_z.setEnabled(True)
-            self.set_focus("z")
-        else:
-            self.form.input_c_x.setEnabled(True)
-            self.form.input_c_y.setEnabled(True)
-            self.form.input_c_z.setEnabled(True)
-            self.set_focus()
-
-    def set_focus(self, key=None):
-        """Set the focus on the widget that receives the key signal."""
-        if key is None or key == "x":
-            self.form.input_c_x.setFocus()
-            self.form.input_c_x.selectAll()
-        elif key == "y":
-            self.form.input_c_y.setFocus()
-            self.form.input_c_y.selectAll()
-        elif key == "z":
-            self.form.input_c_z.setFocus()
-            self.form.input_c_z.selectAll()
-
-    def reject(self):
-        """Execute when clicking the Cancel button or pressing Escape."""
-        _msg(_tr("Aborted:") + " {}".format(_tr(self.name)))
-        self.finish()
-
-    def finish(self):
-        """Finish the command, after accept or reject.
-
-        It finally calls the parent class to execute
-        the delayed functions, and perform cleanup.
-        """
-        # App.ActiveDocument.commitTransaction()
-        Gui.ActiveDocument.resetEdit()
-        # Runs the parent command to complete the call
-        self.source_command.completed()
